Fix compiler warnings
If using a compiler that has "sizeof(long) > sizeof(int)", there were
some warnings. When I started this project, I used the "long" type if
I wanted to have at least 32 bits. At some point it became clear that
the "int" types must have at least 32 bits too, and I started using
preprocessor directives for testing this.

Functions crc32 and crc32c still return "unsigned long" types because
they should be "strictly conforming", according to the C standard. I
think these are alright, but there are a lot of "unsigned long" types
in the tools, e.g. tools/dy-link, although now it would be enough to
use the "unsigned int" types.

However, there are some advantages. Because different compilers are
used and the value of sizeof(long) varies, and everything still seems
to work, it might be less likely that there are unwanted dependencies
on the particular implementation. Hopefully this whole "story" has an
important aspect: at least I _tried_ to take these kind of details
into account and this mindset could be useful in systems programming.

The fact remains that it was a mistake to even consider that I could
build the tools on a platform that has "sizeof(int) == 2". There are
buffers and all kind of things that would have been very complicated
if I had followed that path (assuming that size_t had the same size).

+++ boot/init/fs.c
@@ -175,7 +175,7 @@ int fs_update_config_at(struct b_video_info *vi)
 	static const char *name = "system/config.at";
 	unsigned char config_at[128];
 	const int fd = 0;
-	unsigned crc, val;
+	unsigned long crc, val;
 	size_t size;
 	int r;
